Add method to get a transaction

+++ src/Core/Transactions/Services/IBusinessService.cs
@@ -9,7 +9,7 @@ namespace IDMONEY.IO.Transactions
 {
     public interface IBusinessService
     {
-        Task<BusinessResponse> GetAsync(int id);
+        Task<BusinessResponse> GetAsync(long id);
         Task<SearchBusinessResponse> FindByNameAsync(string name);
         Task<InsertBusinessResponse> CreateAsync(CreateBusinessRequest request);
         Task<SearchBusinessResponse> GetAllAsync();
